% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meSuSie_QC.R
\name{kriging_rss}
\alias{kriging_rss}
\title{LD mismatch detection (Adapted from SuSiE)}
\usage{
kriging_rss(
  z,
  R,
  r_tol = 1e-08,
  s = estimate_s_rss(z, R, r_tol, method = "null-mle")
)
}
\arguments{
\item{z}{A numeric vector of summary statistics.}

\item{R}{A correlation matrix. If the eigen decomposition hasn't been 
performed on R, it will be computed within the function.}

\item{r_tol}{A numeric tolerance level for positive semidefiniteness 
of R. Default is \( 1e-08 \).}

\item{s}{The parameter \( s \) for RSS-based fine-mapping. 
If not provided, it will be estimated using `estimate_s_rss` function 
with the "null-mle" method.}
}
\value{
A list containing:
\itemize{
  \item \code{plot} A ggplot object visualizing the observed vs. expected 
  z scores.
  \item \code{conditional_dist} A data frame containing the summary 
  statistics, conditional means, variances, standardized differences, 
  and log likelihood ratios.
}
}
\description{
This function performs LD mismatch detection on summary statistics and LD
}
