\name{MESuSiE-package}
\alias{MESuSiE-package}
\alias{MESuSiE}
\docType{package}
\title{\packageTitle{MESuSiE}}
\description{\packageDescription{MESuSiE}}
\section{Package Content}{\packageIndices{MESuSiE}}
\author{\packageAuthor{MESuSiE}}
\section{Maintainer}{\packageMaintainer{MESuSiE}}
\keyword{package}
