% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meSuSie_utility.R
\name{meSuSie_get_cs}
\alias{meSuSie_get_cs}
\title{Extract Credible Sets from MESuSiE Results}
\usage{
meSuSie_get_cs(
  res,
  Xcorr,
  coverage = 0.95,
  prior_tol = 1e-09,
  cor_method = cor_method,
  cor_threshold = cor_threshold
)
}
\arguments{
\item{res}{The result object from MESuSiE.}

\item{Xcorr}{A list of correlation matrix used in MESuSiE.}

\item{coverage}{A numeric value indicating the desired coverage for the credible set, default is \(0.95\).}

\item{prior_tol}{A numeric threshold value to determine prior variances, default is \(1e-9\).}

\item{cor_method}{A string indicating which method to use when computing purity, defalt is "min.abs.corr".}

\item{cor_threshold}{A numeric threshold value for purity measurement, default is 0.5.}
}
\value{
A list containing:
\itemize{
  \item \code{cs} The credible sets.
  \item \code{cs_category} The category of each credible set being shared or ancestry specific.
  \item \code{purity} Purity statistics for each credible set.
  \item \code{cs_index} The index for the credible sets.
  \item \code{coverage} The achieved coverage for the credible sets.
  \item \code{requested_coverage} The desired coverage (provided as input).
}
}
\description{
This function processes results from MESuSiE to identify credible sets,
compute their purity and provide other associated statistics.
}
