% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meSuSie_utility.R
\name{MESuSiE_Plot}
\alias{MESuSiE_Plot}
\title{Create GWAS and PIP Plots from MESuSiE Results}
\usage{
MESuSiE_Plot(res, R_mat, summary_data)
}
\arguments{
\item{res}{The result object from MESuSiE.}

\item{R_mat}{A list containing LD (Linkage Disequilibrium) matrices for each ancestry.}

\item{summary_data}{A list of data frames for each ancestry containing at least SNP, Z (Z-scores), 
and possibly POS (position) columns.}
}
\value{
A combined ggplot object of GWAS plots for each ancestry and a PIP plot.
}
\description{
This function generates GWAS (Genome-Wide Association Study) and PIP (Posterior Inclusion Probability)
plots for each ancestry based on the results from MESuSiE.
}
\examples{
\dontrun{
# Assuming `results` is from MESuSiE, `ld_matrices` is the LD matrix list and `data_summary` is the summary data list
MESuSiE_Plot_out <- MESuSiE_Plot(results, ld_matrices, data_summary)
print(MESuSiE_Plot_out)
}

}
