% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meSuSie_R6.R
\name{meSuSieData}
\alias{meSuSieData}
\title{meSuSieData R6 Class}
\description{
The meSuSieData class is used for handling and processing data in the context
of the meSuSie method.
}
\details{
This R6 class is designed to handle and process data relevant to the meSuSie approach.
The class includes methods for initialization and processing XtX and XtY data structures.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{R}}{A list of data.}

\item{\code{Summary_Stat}}{Summary statistics data.}

\item{\code{var_y}}{Variance of y. Default is 1.}

\item{\code{Name_list}}{List of names extracted from `R`.}

\item{\code{N_ancestry}}{Number of ancestry elements in `X`.}

\item{\code{XtX.diag}}{Diagonal elements of XtX.}

\item{\code{XtX_list}}{Processed XtX list.}

\item{\code{Xty_list}}{Processed XtY list.}

\item{\code{N_list}}{List of median values of N from `Summary_Stat`.}

\item{\code{yty_list}}{List of computed yty values.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{R}}{A list of data.}

\item{\code{Summary_Stat}}{Summary statistics data.}

\item{\code{var_y}}{Variance of y. Default is 1.}

\item{\code{Name_list}}{List of names extracted from `R`.}

\item{\code{N_ancestry}}{Number of ancestry elements in `X`.}

\item{\code{XtX.diag}}{Diagonal elements of XtX.}

\item{\code{XtX_list}}{Processed XtX list.}

\item{\code{Xty_list}}{Processed XtY list.}

\item{\code{N_list}}{List of median values of N from `Summary_Stat`.}

\item{\code{yty_list}}{List of computed yty values.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-meSuSieData-new}{\code{meSuSieData$new()}}
\item \href{#method-meSuSieData-XtX_diag}{\code{meSuSieData$XtX_diag()}}
\item \href{#method-meSuSieData-XtX_pro}{\code{meSuSieData$XtX_pro()}}
\item \href{#method-meSuSieData-Xty_pro}{\code{meSuSieData$Xty_pro()}}
\item \href{#method-meSuSieData-clone}{\code{meSuSieData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-meSuSieData-new"></a>}}
\if{latex}{\out{\hypertarget{method-meSuSieData-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{meSuSieData$new(X, Y, var_y = 1)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-meSuSieData-XtX_diag"></a>}}
\if{latex}{\out{\hypertarget{method-meSuSieData-XtX_diag}{}}}
\subsection{Method \code{XtX_diag()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{meSuSieData$XtX_diag(Summary_Stat, Name_list)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-meSuSieData-XtX_pro"></a>}}
\if{latex}{\out{\hypertarget{method-meSuSieData-XtX_pro}{}}}
\subsection{Method \code{XtX_pro()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{meSuSieData$XtX_pro(R, XtX.diag, Name_list)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-meSuSieData-Xty_pro"></a>}}
\if{latex}{\out{\hypertarget{method-meSuSieData-Xty_pro}{}}}
\subsection{Method \code{Xty_pro()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{meSuSieData$Xty_pro(Summary_Stat, XtX.diag, Name_list)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-meSuSieData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-meSuSieData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{meSuSieData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
