% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meSuSie_R6.R
\name{meSuSieObject}
\alias{meSuSieObject}
\title{MESuSiE Object}
\value{
A new `meSuSieObject` object.

A list of computed Xb matrices.

Matrix of computed residuals.

Updated parameters in the `meSuSieObject` object.
}
\description{
An R6 object for handling MESuSiE analysis.

Initialize the MESuSiE object.

Compute the Xb for given b.

Compute the residual.

Update parameters from the SER results.
}
\details{
This object is designed to prepare the MESuSiE analysis.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name_config}}{Character vector for names of the possible configurations.}

\item{\code{column_config}}{List of possible configurations for columns.}

\item{\code{alpha}}{List of alpha matrices for each component.}

\item{\code{mu1}}{List of mu1 matrices for each component.}

\item{\code{mu2}}{List of mu2 matrices for each component.}

\item{\code{EB1}}{List of EB1 matrices for each component.}

\item{\code{EB2}}{List of EB2 matrices for each component.}

\item{\code{Xr}}{List of Xr matrices for each component.}

\item{\code{KL}}{Numeric vector of KL values for each component.}

\item{\code{lbf}}{Numeric vector of lbf values for each component.}

\item{\code{lbf_variable}}{List of lbf values for each variable.}

\item{\code{ELBO}}{Numeric vector of ELBO values for each iteration.}

\item{\code{sigma2}}{Numeric vector of residual variances.}

\item{\code{V}}{List of V matrices for each component.}

\item{\code{pi}}{Numeric vector of prior weights.}

\item{\code{estimate_prior_method}}{Character string specifying the method to estimate prior.}

\item{\code{estimate_residual_variance}}{Logical, if TRUE, the residual variance is estimated.}

\item{\code{L}}{Integer for number of causal components.}

\item{\code{nSNP}}{Integer for number of SNPs.}

\item{\code{nancestry}}{Integer for number of ancestries.}

\item{\code{cs}}{List of credible sets.}

\item{\code{pip}}{Numeric vector of PIP values.}

\item{\code{pip_config}}{Matrix of PIP values for each configuration.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{name_config}}{Character vector for names of the possible configurations.}

\item{\code{column_config}}{List of possible configurations for columns.}

\item{\code{alpha}}{List of alpha matrices for each component.}

\item{\code{mu1}}{List of mu1 matrices for each component.}

\item{\code{mu2}}{List of mu2 matrices for each component.}

\item{\code{EB1}}{List of EB1 matrices for each component.}

\item{\code{EB2}}{List of EB2 matrices for each component.}

\item{\code{Xr}}{List of Xr matrices for each component.}

\item{\code{KL}}{Numeric vector of KL values for each component.}

\item{\code{lbf}}{Numeric vector of lbf values for each component.}

\item{\code{lbf_variable}}{List of lbf values for each variable.}

\item{\code{ELBO}}{Numeric vector of ELBO values for each iteration.}

\item{\code{sigma2}}{Numeric vector of residual variances.}

\item{\code{V}}{List of V matrices for each component.}

\item{\code{pi}}{Numeric vector of prior weights.}

\item{\code{estimate_prior_method}}{Character string specifying the method to estimate prior.}

\item{\code{estimate_residual_variance}}{Logical, if TRUE, the residual variance is estimated.}

\item{\code{L}}{Integer for number of causal components.}

\item{\code{nSNP}}{Integer for number of SNPs.}

\item{\code{nancestry}}{Integer for number of ancestries.}

\item{\code{cs}}{List of credible sets.}

\item{\code{pip}}{Numeric vector of PIP values.}

\item{\code{pip_config}}{Matrix of PIP values for each configuration.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-meSuSieObject-new}{\code{meSuSieObject$new()}}
\item \href{#method-meSuSieObject-compute_Xb}{\code{meSuSieObject$compute_Xb()}}
\item \href{#method-meSuSieObject-compute_residual}{\code{meSuSieObject$compute_residual()}}
\item \href{#method-meSuSieObject-compute_Xr}{\code{meSuSieObject$compute_Xr()}}
\item \href{#method-meSuSieObject-par_update}{\code{meSuSieObject$par_update()}}
\item \href{#method-meSuSieObject-compute_SER_posterior_loglik}{\code{meSuSieObject$compute_SER_posterior_loglik()}}
\item \href{#method-meSuSieObject-compute_KL}{\code{meSuSieObject$compute_KL()}}
\item \href{#method-meSuSieObject-update_residual_variance}{\code{meSuSieObject$update_residual_variance()}}
\item \href{#method-meSuSieObject-get_result}{\code{meSuSieObject$get_result()}}
\item \href{#method-meSuSieObject-mesusie_summary}{\code{meSuSieObject$mesusie_summary()}}
\item \href{#method-meSuSieObject-clone}{\code{meSuSieObject$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-meSuSieObject-new"></a>}}
\if{latex}{\out{\hypertarget{method-meSuSieObject-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{meSuSieObject$new(
  p,
  L,
  N_ancestry,
  residual_variance,
  prior_weights,
  estimate_prior_method,
  estimate_residual_variance,
  max_iter,
  name_vec
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{p}}{Integer, number of SNPs.}

\item{\code{L}}{Integer, number of causal components.}

\item{\code{N_ancestry}}{Integer, number of ancestries.}

\item{\code{residual_variance}}{Numeric vector of initial residual variances.}

\item{\code{prior_weights}}{Numeric vector of prior weights.}

\item{\code{estimate_prior_method}}{Character string, method to estimate prior.}

\item{\code{estimate_residual_variance}}{Logical, if TRUE, estimate the residual variance.}

\item{\code{max_iter}}{Integer, maximum number of iterations.}

\item{\code{name_vec}}{Character vector, names for configurations.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-meSuSieObject-compute_Xb"></a>}}
\if{latex}{\out{\hypertarget{method-meSuSieObject-compute_Xb}{}}}
\subsection{Method \code{compute_Xb()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{meSuSieObject$compute_Xb(meSuSie_Data, b)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{meSuSie_Data}}{List, MESuSiE data object.}

\item{\code{meSuSie_Data}}{List, MESuSiE data object.}

\item{\code{b}}{Matrix, current estimate of b.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-meSuSieObject-compute_residual"></a>}}
\if{latex}{\out{\hypertarget{method-meSuSieObject-compute_residual}{}}}
\subsection{Method \code{compute_residual()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{meSuSieObject$compute_residual(meSuSie_Data, meSuSie_Obj)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{meSuSie_Data}}{List, MESuSiE data object.}

\item{\code{meSuSie_Data}}{List, MESuSiE data object.}

\item{\code{meSuSie_Obj}}{List, current MESuSiE object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-meSuSieObject-compute_Xr"></a>}}
\if{latex}{\out{\hypertarget{method-meSuSieObject-compute_Xr}{}}}
\subsection{Method \code{compute_Xr()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{meSuSieObject$compute_Xr(meSuSie_Data, b1b2, l_index)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{meSuSie_Data}}{List, MESuSiE data object.}

\item{\code{meSuSie_Data}}{List, MESuSiE data object.}

\item{\code{l_index}}{Integer, index for the current component.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-meSuSieObject-par_update"></a>}}
\if{latex}{\out{\hypertarget{method-meSuSieObject-par_update}{}}}
\subsection{Method \code{par_update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{meSuSieObject$par_update(SER_res, l_index)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{SER_res}}{List, results from the SER analysis.}

\item{\code{l_index}}{Integer, index for the current component.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-meSuSieObject-compute_SER_posterior_loglik"></a>}}
\if{latex}{\out{\hypertarget{method-meSuSieObject-compute_SER_posterior_loglik}{}}}
\subsection{Method \code{compute_SER_posterior_loglik()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{meSuSieObject$compute_SER_posterior_loglik(meSuSie_Data, comp_residual, b1b2)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{meSuSie_Data}}{List, MESuSiE data object.}

\item{\code{meSuSie_Data}}{List, MESuSiE data object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-meSuSieObject-compute_KL"></a>}}
\if{latex}{\out{\hypertarget{method-meSuSieObject-compute_KL}{}}}
\subsection{Method \code{compute_KL()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{meSuSieObject$compute_KL(SER_res, value, l_index)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{SER_res}}{List, results from the SER analysis.}

\item{\code{l_index}}{Integer, index for the current component.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-meSuSieObject-update_residual_variance"></a>}}
\if{latex}{\out{\hypertarget{method-meSuSieObject-update_residual_variance}{}}}
\subsection{Method \code{update_residual_variance()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{meSuSieObject$update_residual_variance(meSuSie_Data, niter)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{meSuSie_Data}}{List, MESuSiE data object.}

\item{\code{meSuSie_Data}}{List, MESuSiE data object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-meSuSieObject-get_result"></a>}}
\if{latex}{\out{\hypertarget{method-meSuSieObject-get_result}{}}}
\subsection{Method \code{get_result()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{meSuSieObject$get_result(cs, pip, pip_config)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-meSuSieObject-mesusie_summary"></a>}}
\if{latex}{\out{\hypertarget{method-meSuSieObject-mesusie_summary}{}}}
\subsection{Method \code{mesusie_summary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{meSuSieObject$mesusie_summary(meSuSie_Data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{meSuSie_Data}}{List, MESuSiE data object.}

\item{\code{meSuSie_Data}}{List, MESuSiE data object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-meSuSieObject-clone"></a>}}
\if{latex}{\out{\hypertarget{method-meSuSieObject-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{meSuSieObject$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
