% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meSuSie_QC.R
\name{organize_gwas}
\alias{organize_gwas}
\title{Organize GWAS data from multiple ancestries}
\usage{
organize_gwas(a1, a2, element_names = c("data1", "data2"))
}
\arguments{
\item{a1}{A GWAS dataframe containing SNP information. At a minimum, the dataframe 
must either have columns "SNP", "beta", "se", and "n" or columns "SNP", "Z", and "n". 
Additional columns like "ref", "alt", "chr", and "pos" are recommended but not mandatory.}

\item{a2}{A second GWAS dataframe with SNP details. Similarly to `a1`, it must 
either have columns "SNP", "beta", "se", and "n" or columns "SNP", "Z", and "n". 
Additional columns like "ref", "alt", "chr", and "pos" are recommended but not mandatory.}

\item{element_names}{A character vector of length 2, naming the resulting list elements. Default is c("data1", "data2").}
}
\value{
A list containing two processed and organized genetic dataframes with user-provided element names.
}
\description{
This function processes and organizes GWAS from two ancestries. It ensures that the dataframes have
the necessary columns, checks and warns about any issues, renames columns to a standard format,
and outputs a list containing the processed dataframes.
}
\details{
The function processes the following:
- Checks for required columns.
- Computes Z, Beta, and SE values.
- Checks and warns for strand-ambiguous and multi-allelic SNPs.
- Renames columns to a standard format.
- Warns if SNP sets do not match and subsets to common SNPs.
- Reports NA values in datasets.
- Checks if REF and ALT alleles match and adjusts data if needed.
- Suggests providing chromosome and base pair position for further visualization.
}
