% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meSuSie_QC.R
\name{organize_ld}
\alias{organize_ld}
\title{Organize LD matrices and match them to GWAS data.}
\usage{
organize_ld(ld1, ld2, gwas_data)
}
\arguments{
\item{ld1}{Matrix: The first LD matrix.}

\item{ld2}{Matrix: The second LD matrix.}

\item{gwas_data}{List: The output from the `organize_data` function containing two GWAS datasets.}
}
\value{
A list of two LD matrices with names matching the GWAS datasets from `gwas_data`.
}
\description{
This function checks and organizes two LD matrices to match them with the GWAS data provided by `organize_data`. The LD matrices are adjusted for symmetry if necessary. The SNPs in the column names of the LD matrices should exactly match the SNPs in the GWAS datasets.
}
