% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meSuSie_R6.R
\name{meSuSie_core}
\alias{meSuSie_core}
\title{meSuSie core function}
\usage{
meSuSie_core(
  R_mat_list,
  summary_stat_list,
  L,
  residual_variance = NULL,
  prior_weights = NULL,
  ancestry_weight = NULL,
  optim_method = "optim",
  estimate_residual_variance = F,
  max_iter = 100,
  cor_method = "min.abs.corr",
  cor_threshold = 0.5
)
}
\arguments{
\item{R_mat_list}{A list of length N ancestry with each elment being correlation matrix with dimension p*p, the column name of the correlation matrix should match to the order of SNP name in summary_stat_list}

\item{summary_stat_list}{A list of length N ancestry with each elment being summary statistics. The minimum requirement of summary statistics contains columns of SNP, Beta, Se, Z, and N. The order of the SNP should match the order of the correlation matrix}

\item{L}{Number of effects supposed within the region}
}
\value{
An R6 object with pip, credible sets, and other features of the fine-mapping result.
}
\description{
meSuSie core function
}
