% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meSuSie_R6.R
\name{meSuSieObject}
\alias{meSuSieObject}
\title{MESuSiE Object}
\value{
An R6 object with initials for the MESuSiE analysis.
}
\description{
Prepare MESuSiE object for analysis
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-meSuSieObject-new}{\code{meSuSieObject$new()}}
\item \href{#method-meSuSieObject-compute_Xb}{\code{meSuSieObject$compute_Xb()}}
\item \href{#method-meSuSieObject-compute_residual}{\code{meSuSieObject$compute_residual()}}
\item \href{#method-meSuSieObject-compute_Xr}{\code{meSuSieObject$compute_Xr()}}
\item \href{#method-meSuSieObject-par_update}{\code{meSuSieObject$par_update()}}
\item \href{#method-meSuSieObject-compute_SER_posterior_loglik}{\code{meSuSieObject$compute_SER_posterior_loglik()}}
\item \href{#method-meSuSieObject-compute_KL}{\code{meSuSieObject$compute_KL()}}
\item \href{#method-meSuSieObject-update_residual_variance}{\code{meSuSieObject$update_residual_variance()}}
\item \href{#method-meSuSieObject-get_result}{\code{meSuSieObject$get_result()}}
\item \href{#method-meSuSieObject-mesusie_summary}{\code{meSuSieObject$mesusie_summary()}}
\item \href{#method-meSuSieObject-clone}{\code{meSuSieObject$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-meSuSieObject-new"></a>}}
\if{latex}{\out{\hypertarget{method-meSuSieObject-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{meSuSieObject$new(
  p,
  L,
  N_ancestry,
  residual_variance,
  prior_weights,
  estimate_prior_method,
  estimate_residual_variance,
  max_iter
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-meSuSieObject-compute_Xb"></a>}}
\if{latex}{\out{\hypertarget{method-meSuSieObject-compute_Xb}{}}}
\subsection{Method \code{compute_Xb()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{meSuSieObject$compute_Xb(meSuSie_Data, b)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-meSuSieObject-compute_residual"></a>}}
\if{latex}{\out{\hypertarget{method-meSuSieObject-compute_residual}{}}}
\subsection{Method \code{compute_residual()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{meSuSieObject$compute_residual(meSuSie_Data, meSuSie_Obj)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-meSuSieObject-compute_Xr"></a>}}
\if{latex}{\out{\hypertarget{method-meSuSieObject-compute_Xr}{}}}
\subsection{Method \code{compute_Xr()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{meSuSieObject$compute_Xr(meSuSie_Data, b1b2, l_index)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-meSuSieObject-par_update"></a>}}
\if{latex}{\out{\hypertarget{method-meSuSieObject-par_update}{}}}
\subsection{Method \code{par_update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{meSuSieObject$par_update(SER_res, l_index)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-meSuSieObject-compute_SER_posterior_loglik"></a>}}
\if{latex}{\out{\hypertarget{method-meSuSieObject-compute_SER_posterior_loglik}{}}}
\subsection{Method \code{compute_SER_posterior_loglik()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{meSuSieObject$compute_SER_posterior_loglik(meSuSie_Data, comp_residual, b1b2)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-meSuSieObject-compute_KL"></a>}}
\if{latex}{\out{\hypertarget{method-meSuSieObject-compute_KL}{}}}
\subsection{Method \code{compute_KL()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{meSuSieObject$compute_KL(SER_res, value, l_index)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-meSuSieObject-update_residual_variance"></a>}}
\if{latex}{\out{\hypertarget{method-meSuSieObject-update_residual_variance}{}}}
\subsection{Method \code{update_residual_variance()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{meSuSieObject$update_residual_variance(meSuSie_Data, niter)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-meSuSieObject-get_result"></a>}}
\if{latex}{\out{\hypertarget{method-meSuSieObject-get_result}{}}}
\subsection{Method \code{get_result()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{meSuSieObject$get_result(cs, pip, pip_config)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-meSuSieObject-mesusie_summary"></a>}}
\if{latex}{\out{\hypertarget{method-meSuSieObject-mesusie_summary}{}}}
\subsection{Method \code{mesusie_summary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{meSuSieObject$mesusie_summary(meSuSie_Data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-meSuSieObject-clone"></a>}}
\if{latex}{\out{\hypertarget{method-meSuSieObject-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{meSuSieObject$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
